+++ app/src/main/java/io/github/ferreol/seteditplus/Utils/Shortcut/ShortcutEditRecyclerRowMoveCallback.java
@@ -1,4 +1,4 @@
-package io.github.ferreol.seteditplus.Utils;
+package io.github.ferreol.seteditplus.Utils.Shortcut;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
