+++ app/src/main/java/io/github/ferreol/seteditplus/Utils/Shortcut/ShortcutEditRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package io.github.ferreol.seteditplus.Utils;
+package io.github.ferreol.seteditplus.Utils.Shortcut;
 
 import android.graphics.Color;
 import android.view.LayoutInflater;
@@ -7,7 +7,6 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
-import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.Collections;
