+++ app/src/main/java/io/github/ferreol/seteditplus/Utils/Shortcut/ShortcutIcons.java
@@ -1,4 +1,4 @@
-package io.github.ferreol.seteditplus.Utils;
+package io.github.ferreol.seteditplus.Utils.Shortcut;
 
 import android.app.AlertDialog;
 import android.content.Context;
